Corrected CfgSounds. Moved from maps to core.
