+++ A3A/addons/core/Sounds/Misc/StrikeSound.ogg
