+++ A3A/addons/maps/Antistasi_cam_lao_nam.cam_lao_nam/description.ext
@@ -12,11 +12,6 @@ class CfgNotifications
     #include "..\MissionDescription\CfgNotificationsContents.hpp"
 };
 
-class CfgSounds
-{
-    #include "..\MissionDescription\CfgSoundsContents.hpp"
-};
-
 class CfgDebriefing {
     #include "..\MissionDescription\CfgDebriefingContents.hpp"
 	class End1
