+++ A3A/addons/maps/Antistasi_chernarus_winter.chernarus_winter/description.ext
@@ -12,11 +12,6 @@ class CfgNotifications
     #include "..\missionDescription\CfgNotificationsContents.hpp"
 };
 
-class CfgSounds
-{
-    #include "..\missionDescription\CfgSoundsContents.hpp"
-};
-
 class CfgDebriefing {
     #include "..\missionDescription\CfgDebriefingContents.hpp"
 	class End1
